requirements.txt
flask
flask-cors
supabase
python-dotenv
requests
numpy
pandas
scikit-learn
# Hardware libraries (install only on Pi)
# Adafruit_DHT
# RPi.GPIO
# spidev
